object_detection/requirements.txt
scipy==1.10.1
numpy==1.24.4
opencv-python==4.10.0.84
onnx==1.17.0
onnxruntime==1.16.3
clearblade
python-dotenv
lapx
ultralytics